Initial Refactor including buildParameters and initializeContext
